pyproject.toml
[project]
name = "pi-radio-alerts"
version = "0.1.0"
description = "Quick Call II alert monitor for a Raspberry Pi 5 with RTL-SDR and RGB matrix output."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26,<3",
  "Pillow>=10.0,<12",
]

[project.scripts]
pi-radio-alerts = "pi_radio_alerts.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

